Created endpoints and test files
Created the API endpoints for Weddings, Guests, and Expenses.

Weddings, and Guests I was able to make succesful GET requests. I still need to test POST, DELETE, and PATCH.

Expense GET request received an error which I need to investigate. Couldn't see if any of the test files worked because of password issue.